+++ src/components/Header.js
@@ -1,0 +1,29 @@
+const Header = (/*html*/) => {
+  const loggedIn = isLoggedIn();
+
+  function isLoggedIn() {
+    const userData = JSON.parse(localStorage.getItem("user"));
+
+    return userData !== null;
+  }
+
+  return `
+    <div class="max-w-md w-full">
+        <header class="bg-blue-600 text-white p-4 sticky top-0">
+          <h1 class="text-2xl font-bold">항해플러스</h1>
+        </header>
+        <nav class="bg-white shadow-md p-2 sticky top-14">
+          <ul class="flex justify-around">
+            <li><a href="/" class="text-blue-600">홈</a></li>
+            <li>
+              ${loggedIn ? `<a href="/profile" class="text-gray-600">프로필</a>` : ``}
+            </li>
+            <li>
+              ${loggedIn ? `<a href="/login" id='logout' class="text-gray-600">로그아웃</a>` : `<a href="/login" class="text-gray-600">로그인</a>`}
+            </li>
+          </ul>
+        </nav>
+  `;
+};
+
+export default Header;
